requirements.txt
streamlit>=1.28.0
pokerkit>=1.2.8
treys>=0.2.2
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.15.0
umap-learn>=0.5.3